Add Markdownlint workflow for linting Markdown files
This commit adds a Markdownlint workflow in `.github/workflows/mardownlint.yml` to perform linting on Markdown files in the `docs` directory. The workflow runs on both push and pull request events for the `main` and `develop` branches. It utilizes the `actions/checkout@v3` action to clone the repository and the `articulate/actions-markdownlint@v1` action to perform the linting. The configuration file used is `markdownlint-config.json` and it lints all Markdown files except those in the `node_modules` directory. The version of Markdownlint used is `0.28.1`.