feat: Add .editorconfig file with code formatting rules (#1)
* :truck: Transfer images to docs

* feat: Add .editorconfig file with code formatting rules

- Adds a new file `.editorconfig` with standard code formatting rules using UTF-8 encoding, inserting final newline, and trimming trailing whitespace.
- Defines indentation styles and sizes for different file types to ensure consistent formatting.
- Uses `space` indentation for most file types except for specific ones like Makefile, go files, and Markdown files which use `tab` indentation and have different indentation sizes.
- Sets `indent_size` to 4 for Dockerfile.
- Overall, this commit enhances code consistency and readability.

Refactor: Update README.md with detailed features and specifications

- Updates the project overview and adds extensive descriptions for each feature and functionality of the GeekBot Telegram bot.
- Includes details on new member verification, moderation automation, reward system, point redemption, and community rule enforcement.
- Provides visual representation of community merch in the README.
- Adds a section on the technical specifications, stack, and tooling used in the development of the GeekBot.
- Includes a list of study materials for learning Go.

Note: The README file was not properly formatted, but it has been fixed in this commit.

* feat: Add initial root command implementation

This commit adds the initial implementation of the root command, which represents the base command when called without any subcommands. The root command has a short description of the application and a longer description that provides examples and usage of the application. It also defines a "toggle" flag as a local flag for the root command. Additionally, this commit includes the addition of the go.mod and go.sum files, which specify the Go module and its required indirect dependencies. Finally, the main.go file is modified to import and execute the root command.

* Add Universal Linux dev container config

This commit adds the configuration file for the Universal Linux dev container. It includes references to various features such as kubectl, Helm, Minikube, Node.js, Vim, and k3d. Uncommented lines provide additional customization options like postCreateCommand and the ability to connect as root. For more details, please refer to the README documentation.